Refactor and consolidate MST processing and network generation
- Moved MST processing functions from `MST_processing.py` and `simple_MST.py` into a new file `minimal_spanning_tree.py`.
- Consolidated network generation functions from `simple_MST.py` into `net_generation.py`.
- Updated import paths in `import_and_create_layers.py` to reflect new structure.
- Removed deprecated files `MST_processing.py` and `simple_MST.py`.
- Enhanced the `generate_network` function to utilize the new structure for generating heat networks.
- Adjusted the logic for creating and adjusting segments to follow street lines.
- Improved code organization for better maintainability and readability.
