+++ examples/data/Wärmenetz/Variante 1/Wärmenetz/Rücklauf.geojson
@@ -8,27 +8,27 @@
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499809.349299236142542, 5666252.651789142750204 ], [ 499811.634540085040499, 5666280.477533169090748 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499809.349299236142542, 5666252.651789142750204 ], [ 499782.426409600244369, 5666198.836034059524536 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499782.426409600244369, 5666198.836034059524536 ], [ 499782.424584617256187, 5666198.833520421758294 ] ] } },
-{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499812.294332949153613, 5666307.668479453772306 ], [ 499812.11814906593645, 5666317.808907736092806 ] ] } },
-{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499812.294332949153613, 5666307.668479453772306 ], [ 499812.489200502284802, 5666296.452695065177977 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499812.15308640699368, 5666414.18343029730022 ], [ 499823.899424955365248, 5666425.339984565973282 ] ] } },
-{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499812.15308640699368, 5666414.18343029730022 ], [ 499795.736449335061479, 5666413.284210711717606 ] ] } },
-{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499781.741014994273428, 5666440.257240979932249 ], [ 499791.537854722933844, 5666421.376051007770002 ] ] } },
+{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499812.15308640699368, 5666414.18343029730022 ], [ 499795.095732904330362, 5666414.519046534784138 ] ] } },
+{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499781.741014994273428, 5666440.257240979932249 ], [ 499789.261409403115977, 5666425.7633838346228 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499781.741014994273428, 5666440.257240979932249 ], [ 499767.976320928893983, 5666466.785571620799601 ] ] } },
+{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499812.489200502284802, 5666296.452695065177977 ], [ 499812.228235343820415, 5666311.472788989543915 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499812.489200502284802, 5666296.452695065177977 ], [ 499811.634540085040499, 5666280.477533169090748 ] ] } },
-{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499812.11814906593645, 5666317.808907736092806 ], [ 499824.749122653796803, 5666325.074632585979998 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499730.333925593295135, 5666503.816372456029058 ], [ 499733.486116384039633, 5666503.128313224762678 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499730.333925593295135, 5666503.816372456029058 ], [ 499666.264326982316561, 5666518.030200897715986 ] ] } },
+{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499812.07416039344389, 5666320.340716948732734 ], [ 499812.228235343820415, 5666311.472788989543915 ] ] } },
+{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499812.07416039344389, 5666320.340716948732734 ], [ 499824.749122653796803, 5666325.074632585979998 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499666.264326982316561, 5666518.030200897715986 ], [ 499631.246108009654563, 5666520.504589550197124 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499875.099407144938596, 5666461.546726420521736 ], [ 499897.088487282046117, 5666474.221831605769694 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499875.099407144938596, 5666461.546726420521736 ], [ 499840.663302131637465, 5666437.159727067686617 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499823.899424955365248, 5666425.339984565973282 ], [ 499840.663302131637465, 5666437.159727067686617 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499782.422759634209797, 5666198.831006784923375 ], [ 499782.424584617256187, 5666198.833520421758294 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499856.314965481113177, 5666313.174454231746495 ], [ 499841.1039219412487, 5666318.908948066644371 ] ] } },
+{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499789.261409403115977, 5666425.7633838346228 ], [ 499795.095732904330362, 5666414.519046534784138 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499767.976320928893983, 5666466.785571620799601 ], [ 499761.584232726832852, 5666489.200370393693447 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499631.246108009654563, 5666520.504589550197124 ], [ 499597.462798655731604, 5666511.515230836346745 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499761.584232726832852, 5666489.200370393693447 ], [ 499750.358384107530583, 5666499.445439839735627 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499750.358384107530583, 5666499.445439839735627 ], [ 499736.638307174725924, 5666502.440253992564976 ] ] } },
-{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499795.736449335061479, 5666413.284210711717606 ], [ 499791.537854722933844, 5666421.376051007770002 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499824.749122653796803, 5666325.074632585979998 ], [ 499841.1039219412487, 5666318.908948066644371 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499733.486116384039633, 5666503.128313224762678 ], [ 499736.638307174725924, 5666502.440253992564976 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499829.544806176505517, 5666164.618855013512075 ], [ 499782.422759634209797, 5666198.831006784923375 ] ] } },
