+++ examples/data/Wärmenetz/Variante 1/Wärmenetz/Vorlauf.geojson
@@ -8,27 +8,27 @@
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499808.849299236142542, 5666252.651789142750204 ], [ 499811.134540085040499, 5666280.477533169090748 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499808.849299236142542, 5666252.651789142750204 ], [ 499781.926409600244369, 5666198.836034059524536 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499781.926409600244369, 5666198.836034059524536 ], [ 499781.924584617256187, 5666198.833520421758294 ] ] } },
-{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499811.794332949153613, 5666307.668479453772306 ], [ 499811.61814906593645, 5666317.808907736092806 ] ] } },
-{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499811.794332949153613, 5666307.668479453772306 ], [ 499811.989200502284802, 5666296.452695065177977 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499811.65308640699368, 5666414.18343029730022 ], [ 499823.399424955365248, 5666425.339984565973282 ] ] } },
-{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499811.65308640699368, 5666414.18343029730022 ], [ 499795.236449335061479, 5666413.284210711717606 ] ] } },
-{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499781.241014994273428, 5666440.257240979932249 ], [ 499791.037854722933844, 5666421.376051007770002 ] ] } },
+{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499811.65308640699368, 5666414.18343029730022 ], [ 499794.595732904330362, 5666414.519046534784138 ] ] } },
+{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499781.241014994273428, 5666440.257240979932249 ], [ 499788.761409403115977, 5666425.7633838346228 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499781.241014994273428, 5666440.257240979932249 ], [ 499767.476320928893983, 5666466.785571620799601 ] ] } },
+{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499811.989200502284802, 5666296.452695065177977 ], [ 499811.728235343820415, 5666311.472788989543915 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499811.989200502284802, 5666296.452695065177977 ], [ 499811.134540085040499, 5666280.477533169090748 ] ] } },
-{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499811.61814906593645, 5666317.808907736092806 ], [ 499824.249122653796803, 5666325.074632585979998 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499729.833925593295135, 5666503.816372456029058 ], [ 499732.986116384039633, 5666503.128313224762678 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499729.833925593295135, 5666503.816372456029058 ], [ 499665.764326982316561, 5666518.030200897715986 ] ] } },
+{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499811.57416039344389, 5666320.340716948732734 ], [ 499811.728235343820415, 5666311.472788989543915 ] ] } },
+{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499811.57416039344389, 5666320.340716948732734 ], [ 499824.249122653796803, 5666325.074632585979998 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499665.764326982316561, 5666518.030200897715986 ], [ 499630.746108009654563, 5666520.504589550197124 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499874.599407144938596, 5666461.546726420521736 ], [ 499896.588487282046117, 5666474.221831605769694 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499874.599407144938596, 5666461.546726420521736 ], [ 499840.163302131637465, 5666437.159727067686617 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499823.399424955365248, 5666425.339984565973282 ], [ 499840.163302131637465, 5666437.159727067686617 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499781.922759634209797, 5666198.831006784923375 ], [ 499781.924584617256187, 5666198.833520421758294 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499855.814965481113177, 5666313.174454231746495 ], [ 499840.6039219412487, 5666318.908948066644371 ] ] } },
+{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499788.761409403115977, 5666425.7633838346228 ], [ 499794.595732904330362, 5666414.519046534784138 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499767.476320928893983, 5666466.785571620799601 ], [ 499761.084232726832852, 5666489.200370393693447 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499630.746108009654563, 5666520.504589550197124 ], [ 499596.962798655731604, 5666511.515230836346745 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499761.084232726832852, 5666489.200370393693447 ], [ 499749.858384107530583, 5666499.445439839735627 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499749.858384107530583, 5666499.445439839735627 ], [ 499736.138307174725924, 5666502.440253992564976 ] ] } },
-{ "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499795.236449335061479, 5666413.284210711717606 ], [ 499791.037854722933844, 5666421.376051007770002 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499824.249122653796803, 5666325.074632585979998 ], [ 499840.6039219412487, 5666318.908948066644371 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499732.986116384039633, 5666503.128313224762678 ], [ 499736.138307174725924, 5666502.440253992564976 ] ] } },
 { "type": "Feature", "properties": { }, "geometry": { "type": "LineString", "coordinates": [ [ 499829.044806176505517, 5666164.618855013512075 ], [ 499781.922759634209797, 5666198.831006784923375 ] ] } },
